backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
pymysql==1.1.1
sqlalchemy==2.0.32
werkzeug==3.0.4
requests==2.32.3
python-dotenv==1.0.1
gunicorn==21.2.0

torch>=2.2.0
diffusers>=0.30.0
Pillow>=10.3.0
google-generativeai>=0.7.2

